pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opp"
version = "0.1.0"
description = "Open Provenance Protocol — Shazam for AI-generated images"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "OPP Contributors" },
]
keywords = ["ai", "provenance", "image", "hashing", "protocol"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "Pillow>=10.0.0",
    "imagehash>=4.3.0",
    "open-clip-torch>=2.24.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.0",
    "qdrant-client>=1.7.0",
    "python-multipart>=0.0.6",
    "rich>=13.0.0",
    "pdqhash>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.28.0",
]

[project.scripts]
opp = "opp.cli:cli"

[project.urls]
Homepage = "https://github.com/opp-protocol/opp"
Documentation = "https://github.com/opp-protocol/opp"
Repository = "https://github.com/opp-protocol/opp"

[tool.setuptools.packages.find]
include = ["opp*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
